Fix crash if qstat is missing when querying a server, fix #179
In the 'while (current)' loop in stat_servers_input_callback(), the
string list usually has a NULL string (input had newline with no text)
at the end of the server entry. However, there was no NULL string for
the error message when qstat command is not found resulting in a crash.

An alternate way to fix this would be to add a newline to error_msg in
start_qstat().