Make possible to do a portable build of XQF
Make possible to do a portable build of XQF.

- USE_RELATIVE_PREFIX: Make the executable looks around for its own files.
- USE_FHS: Install files following the File Hierarchy Standard.

This is how one can do a portable installation where the binaries
are stored at the root of the install folder and the required
data is searched next to them at run time:

cmake -DUSE_RELATIVE_PREFIX=ON -DUSE_FHS=OFF

It makes possible to build and distribute compiled binaries in a tarball.

This relies on some code under the BSD license I copied from the
Dæmon engine¹. I place the glue around under the CC0 1.0 license
(public domain like).

¹ https://github.com/DaemonEngine/Daemon